Use size_t for ethernet package length

+++ core/net/uDTN/convergence_layer_udp.c
@@ -163,7 +163,7 @@ static void convergence_layer_udp_discovery_thread(void *arg)
  * @param length
  * @return
  */
-int convergence_layer_udp_send_discovery(const uint8_t* const payload, const uint8_t length)
+int convergence_layer_udp_send_discovery(const uint8_t* const payload, const size_t length)
 {
 	return convergence_layer_udp_send(discovery_conn, &udp_mcast_addr, CL_UDP_DISCOVERY_PORT, payload, length, NULL, 0);
 }
@@ -204,8 +204,8 @@ static void convergence_layer_udp_bundle_thread(void *arg)
 }
 
 
-int convergence_layer_udp_send_data(const ip_addr_t* const addr, const uint8_t* const payload, const uint8_t length,
-									const uint8_t* const payload2, const uint8_t length2)
+int convergence_layer_udp_send_data(const ip_addr_t* const addr, const uint8_t* const payload, const size_t length,
+									const uint8_t* const payload2, const size_t length2)
 {
 	return convergence_layer_udp_send(bundle_conn, addr, CL_UDP_BUNDLE_PORT, payload, length, payload2, length2);
 }
